removing the separate class for getProductAttributes
