+++ src/lib/api/product.ts
@@ -10,6 +10,19 @@ export class ProductsApi {
 		this.fetch = fetch;
 	}
 
+	async getProductAttributes(barcode: string): Promise<ProductAttribute[]> {
+		const url = `${API_HOST}/api/v2/product/${barcode}?fields=product_name,code,attribute_groups_en`;
+
+		const res = await this.fetch(url);
+
+		if (!res.ok) {
+			throw new Error(`Failed to fetch product attributes for barcode: ${barcode}`);
+		}
+
+		const data = await res.json();
+		return data.product?.attribute_groups_en || [];
+	}
+
 	async getProduct<T extends Array<keyof Product>>(
 		barcode: string,
 		{ fields }: { fields: T } = { fields: ['all', 'knowledge_panels'] as T }
@@ -161,6 +174,23 @@ export type ProductSearch<T = Product> = {
 	products: T[];
 	skip: number;
 };
+
+export type Attribute = {
+	id: string;
+	name: string;
+	grade: string;
+	title: string;
+	description_short?: string;
+	icon_url?: string;
+};
+
+export type ProductAttribute = {
+	id: string;
+	name: string;
+	attributes: Attribute[];
+};
+
+export type ProductAttributes = ProductAttribute[];
 
 type LangIngredient = `ingredients_text_${string}`;
 type LangProduct = `product_name_${string}`;
